rename mocks package to use houston_mocks
