+++ version/validate_compatibility_test.go
@@ -6,7 +6,7 @@ import (
 
 	testUtil "github.com/astronomer/astro-cli/pkg/testing"
 
-	"github.com/astronomer/astro-cli/airflow/mocks"
+	mocks "github.com/astronomer/astro-cli/houston/mocks"
 	"github.com/stretchr/testify/assert"
 )
